fix(Operator): 정수 더하기에서 실수 더하기로 수정

+++ src/main/java/calculator/domain/Operator.java
@@ -10,7 +10,7 @@ public enum Operator {
         public Number apply(Numbers numbers) {
             return new Number(
                     numbers.getNumbers().stream()
-                            .mapToInt(Number::getValue)
+                            .mapToDouble(Number::getValue)
                             .sum()
             );
         }
